Refactor file system and naming conventions + remove unneeded files
