+++ src/Services/service.typings.ts
@@ -1,0 +1,1 @@
+export type Period = 'Q1' | 'Q2' | 'Q3' | 'Q4' | 'FY';
