+++ src/Services/statement/statement.typings.ts
@@ -1,0 +1,148 @@
+import { Period } from "services/service.typings";
+
+export interface Statements {
+    incomeStatements: IncomeStatement[],
+    balanceSheets: BalanceSheet[],
+    cashFlowStatements: CashFlowStatement[]
+}
+
+export interface IncomeStatement {
+    cik: string,
+    date: Date,
+    symbol: string | null,
+    reportedCurrency: string | null,
+    fillingDate: Date | null,
+    acceptedDate: string | null,
+    calendarYear: string | null,
+    period: Period,
+    revenue: number,
+    costOfRevenue: number | null,
+    grossProfit: number | null,
+    grossProfitRatio: number | null,
+    researchAndDevelopmentExpenses: number | null,
+    generalAndAdministrativeExpenses: number | null,
+    sellingAndMarketingExpenses: number | null,
+    sellingGeneralAndAdministrativeExpenses: number | null,
+    otherExpenses: number | null,
+    operatingExpenses: number | null,
+    costAndExpenses: number | null,
+    interestIncome: number | null,
+    interestExpense: number | null,
+    depreciationAndAmortization: number | null,
+    ebitda: number | null,
+    ebitdaratio: number | null,
+    operatingIncome: number,
+    operatingIncomeRatio: number | null,
+    totalOtherIncomeExpensesNet: number | null,
+    incomeBeforeTax: number | null,
+    incomeBeforeTaxRatio: number | null,
+    incomeTaxExpense: number,
+    netIncome: number,
+    netIncomeRatio: number | null,
+    eps: number,
+    epsdiluted: number | null,
+    weightedAverageShsOut: number,
+    weightedAverageShsOutDil: number | null,
+    link: string | null,
+    finalLink: string  | null
+}
+
+export interface BalanceSheet {
+    cik: string,
+    date: Date,
+    symbol: string,
+    reportedCurrency: string | null,
+    fillingDate: Date | null,
+    acceptedDate: string | null,
+    calendarYear: string | null,
+    period: Period,
+    cashAndCashEquivalents: number | null,
+    shortTermInvestments: number | null,
+    cashAndShortTermInvestments: number | null,
+    netReceivables: number | null,
+    inventory: number | null,
+    otherCurrentAssets: number | null,
+    totalCurrentAssets: number | null,
+    propertyPlantEquipmentNet: number | null,
+    goodwill: number | null,
+    intangibleAssets: number | null,
+    goodwillAndIntangibleAssets: number | null,
+    longTermInvestments: number | null,
+    taxAssets: number | null,
+    otherNonCurrentAssets: number | null,
+    totalNonCurrentAssets: number | null,
+    otherAssets: number | null,
+    totalAssets: number | null,
+    accountPayables: number | null,
+    shortTermDebt: number | null,
+    taxPayables: number | null,
+    deferredRevenue: number | null,
+    otherCurrentLiabilities: number | null,
+    totalCurrentLiabilities: number | null,
+    longTermDebt: number,
+    deferredRevenueNonCurrent: number | null,
+    deferredTaxLiabilitiesNonCurrent: number | null,
+    otherNonCurrentLiabilities: number | null,
+    totalNonCurrentLiabilities: number | null,
+    otherLiabilities: number | null,
+    capitalLeaseObligations: number | null,
+    totalLiabilities: number | null,
+    preferredStock: number | null,
+    commonStock: number | null,
+    retainedEarnings: number | null,
+    accumulatedOtherComprehensiveIncomeLoss: number | null,
+    othertotalStockholdersEquity: number | null,
+    totalStockholdersEquity: number,
+    totalEquity: number,
+    totalLiabilitiesAndStockholdersEquity: number | null,
+    minorityInterest: number | null,
+    totalLiabilitiesAndTotalEquity: number | null,
+    totalInvestments: number | null,
+    totalDebt: number | null,
+    netDebt: number,
+    link: string | null,
+    finalLink: string | null
+}
+
+export interface CashFlowStatement {
+    cik: string,
+    date: Date,
+    symbol: string,
+    reportedCurrency: | null,
+    fillingDate: Date | null,
+    acceptedDate: | null,
+    calendarYear: | null,
+    period: Period,
+    netIncome: number,
+    depreciationAndAmortization: number,
+    deferredIncomeTax: number,
+    stockBasedCompensation: number,
+    changeInWorkingCapital: number,
+    accountsReceivables: number,
+    inventory: number,
+    accountsPayables: number,
+    otherWorkingCapital: number,
+    otherNonCashItems: number,
+    netCashProvidedByOperatingActivities: number,
+    investmentsInPropertyPlantAndEquipment: number,
+    acquisitionsNet: number,
+    purchasesOfInvestments: number,
+    salesMaturitiesOfInvestments: number,
+    otherInvestingActivites: number,
+    netCashUsedForInvestingActivites: number,
+    debtRepayment: number,
+    commonStockIssued: number,
+    commonStockRepurchased: number,
+    dividendsPaid: number,
+    otherFinancingActivites: number,
+    netCashUsedProvidedByFinancingActivities: number,
+    effectOfForexChangesOnCash: number,
+    netChangeInCash: number,
+    cashAtEndOfPeriod: number,
+    cashAtBeginningOfPeriod: number,
+    operatingCashFlow: number,
+    capitalExpenditure: number,
+    freeCashFlow: number,
+    link: string,
+    finalLink: string 
+}
