+++ src/Services/sticker-price/helpers/calculator/functions/AbstractFunction.ts
@@ -1,0 +1,9 @@
+import StickerPriceData, { PeriodicData } from "../../../../../services/sticker-price/sticker-price.typings";
+
+abstract class AbstractFunction {
+
+    abstract calculate(stickerPriceData: StickerPriceData): Promise<PeriodicData[]>;
+
+}
+
+export default AbstractFunction;
