+++ src/Services/sticker-price/helpers/calculator/outputs/AbstractOutput.ts
