--- src/resources/entities/statements/balance-sheet.ts
@@ -1,60 +1,0 @@
-import { Period } from "../models/PeriodicData";
-
-export default interface BalanceSheet {
-
-    cik: string,
-    date: Date,
-    symbol: string,
-    reportedCurrency: string | null,
-    fillingDate: Date | null,
-    acceptedDate: string | null,
-    calendarYear: string | null,
-    period: Period,
-    cashAndCashEquivalents: number | null,
-    shortTermInvestments: number | null,
-    cashAndShortTermInvestments: number | null,
-    netReceivables: number | null,
-    inventory: number | null,
-    otherCurrentAssets: number | null,
-    totalCurrentAssets: number | null,
-    propertyPlantEquipmentNet: number | null,
-    goodwill: number | null,
-    intangibleAssets: number | null,
-    goodwillAndIntangibleAssets: number | null,
-    longTermInvestments: number | null,
-    taxAssets: number | null,
-    otherNonCurrentAssets: number | null,
-    totalNonCurrentAssets: number | null,
-    otherAssets: number | null,
-    totalAssets: number | null,
-    accountPayables: number | null,
-    shortTermDebt: number | null,
-    taxPayables: number | null,
-    deferredRevenue: number | null,
-    otherCurrentLiabilities: number | null,
-    totalCurrentLiabilities: number | null,
-    longTermDebt: number,
-    deferredRevenueNonCurrent: number | null,
-    deferredTaxLiabilitiesNonCurrent: number | null,
-    otherNonCurrentLiabilities: number | null,
-    totalNonCurrentLiabilities: number | null,
-    otherLiabilities: number | null,
-    capitalLeaseObligations: number | null,
-    totalLiabilities: number | null,
-    preferredStock: number | null,
-    commonStock: number | null,
-    retainedEarnings: number | null,
-    accumulatedOtherComprehensiveIncomeLoss: number | null,
-    othertotalStockholdersEquity: number | null,
-    totalStockholdersEquity: number,
-    totalEquity: number,
-    totalLiabilitiesAndStockholdersEquity: number | null,
-    minorityInterest: number | null,
-    totalLiabilitiesAndTotalEquity: number | null,
-    totalInvestments: number | null,
-    totalDebt: number | null,
-    netDebt: number,
-    link: string | null,
-    finalLink: string | null
-
-}
