--- src/resources/entities/statements/cashflow-statement.ts
@@ -1,46 +1,0 @@
-import { Period } from "../models/PeriodicData";
-
-export default interface CashFlowStatement {
-
-    cik: string,
-    date: Date,
-    symbol: string,
-    reportedCurrency: | null,
-    fillingDate: Date | null,
-    acceptedDate: | null,
-    calendarYear: | null,
-    period: Period,
-    netIncome: number,
-    depreciationAndAmortization: number,
-    deferredIncomeTax: number,
-    stockBasedCompensation: number,
-    changeInWorkingCapital: number,
-    accountsReceivables: number,
-    inventory: number,
-    accountsPayables: number,
-    otherWorkingCapital: number,
-    otherNonCashItems: number,
-    netCashProvidedByOperatingActivities: number,
-    investmentsInPropertyPlantAndEquipment: number,
-    acquisitionsNet: number,
-    purchasesOfInvestments: number,
-    salesMaturitiesOfInvestments: number,
-    otherInvestingActivites: number,
-    netCashUsedForInvestingActivites: number,
-    debtRepayment: number,
-    commonStockIssued: number,
-    commonStockRepurchased: number,
-    dividendsPaid: number,
-    otherFinancingActivites: number,
-    netCashUsedProvidedByFinancingActivities: number,
-    effectOfForexChangesOnCash: number,
-    netChangeInCash: number,
-    cashAtEndOfPeriod: number,
-    cashAtBeginningOfPeriod: number,
-    operatingCashFlow: number,
-    capitalExpenditure: number,
-    freeCashFlow: number,
-    link: string,
-    finalLink: string 
-
-}
