--- src/resources/entities/statements/income-statement.ts
@@ -1,44 +1,0 @@
-import { Period } from "../models/PeriodicData";
-
-export default interface IncomeStatement {
-
-    cik: string,
-    date: Date,
-    symbol: string | null,
-    reportedCurrency: string | null,
-    fillingDate: Date | null,
-    acceptedDate: string | null,
-    calendarYear: string | null,
-    period: Period,
-    revenue: number,
-    costOfRevenue: number | null,
-    grossProfit: number | null,
-    grossProfitRatio: number | null,
-    researchAndDevelopmentExpenses: number | null,
-    generalAndAdministrativeExpenses: number | null,
-    sellingAndMarketingExpenses: number | null,
-    sellingGeneralAndAdministrativeExpenses: number | null,
-    otherExpenses: number | null,
-    operatingExpenses: number | null,
-    costAndExpenses: number | null,
-    interestIncome: number | null,
-    interestExpense: number | null,
-    depreciationAndAmortization: number | null,
-    ebitda: number | null,
-    ebitdaratio: number | null,
-    operatingIncome: number,
-    operatingIncomeRatio: number | null,
-    totalOtherIncomeExpensesNet: number | null,
-    incomeBeforeTax: number | null,
-    incomeBeforeTaxRatio: number | null,
-    incomeTaxExpense: number,
-    netIncome: number,
-    netIncomeRatio: number | null,
-    eps: number,
-    epsdiluted: number | null,
-    weightedAverageShsOut: number,
-    weightedAverageShsOutDil: number | null,
-    link: string | null,
-    finalLink: string  | null
-    
-}
