+++ src/utils/initializeEnv.ts
@@ -1,11 +1,10 @@
 import { cleanEnv, port } from "envalid";
 import  { GetParametersByPathResult, SSM } from '@aws-sdk/client-ssm';
-import CONSTANTS from "../Services/ServiceConstants";
 import EnvInitializationException from "./exceptions/EnvInitializationException";
 import { AWSError } from "sqs-consumer";
+import CONSTANTS from "@/resources/resource.contants";
 
 async function initializeEnv(): Promise<void> {
-    const filename = 'parameters.json';
     return setEnvParameters().then(() => {
         cleanEnv(process.env, {
             service_port: port({ default: 3000 })
